+++ pycvvdp/__init__.py
@@ -1,7 +1,6 @@
 from pycvvdp.cvvdp_metric import cvvdp
-from pycvvdp.cvvdp_nn_metric import cvvdp_nn
 from pycvvdp.pupsnr import pu_psnr_y, pu_psnr_rgb2020, psnr_rgb
-from pycvvdp.video_source_file import video_source_file, video_source_video_file, load_image_as_array
+from pycvvdp.video_source_file import video_source_file, video_source_video_file, load_image_as_array, video_source_temp_resample_file
 from pycvvdp.display_model import vvdp_display_photometry, vvdp_display_photo_eotf, vvdp_display_geometry
 from pycvvdp.video_source_yuv import video_source_yuv_file
 from pycvvdp.cvvdp_ml_metric import cvvdp_ml_dis_TR, cvvdp_ml, cvvdp_ml_att, cvvdp_ml_transformer